LayoutTests/editing/selection/move-by-word-visually-expected.txt


======== Move By Word ====
Test 1, LTR:
Move right by one word
"abc def hij opq rst"[0, 4, 8, 12, 16]
Move left by one word
"abc def hij opq rst"[19, 16, 12, 8, 4, 0]
Test 2, LTR:
Move right by one word
"abc def hij opq rst"[0, 4, 8, 12, 16]
Move left by one word
"abc def hij opq rst"[19, 16, 12, 8, 4, 0]
Test 3, RTL:
Move left by one word
"abc def hij opq rst"[0, 15, 11, 7, 3]
Move right by one word
"abc def hij opq rst"[19, 3, 7, 11, 15, 0]
Test 4, LTR:
Move right by one word
"ZZZ QQQ BBB CCC XXX"[0, 15, 11, 7, 3]
Move left by one word
"ZZZ QQQ BBB CCC XXX"[19, 3, 7, 11, 15, 0]
Test 5, RTL:
Move left by one word
"ZZZ QQQ BBB CCC XXX"[0, 4, 8, 12, 16]
Move right by one word
"ZZZ QQQ BBB CCC XXX"[19, 16, 12, 8, 4, 0]
Test 6, LTR:
Move right by one word
"abc def hij AAA BBB WWW opr uvw xyz"[0, 4, 8, 12, 19, 15, 24, 28, 32]
Move left by one word
"abc def hij AAA BBB WWW opr uvw xyz"[35, 32, 28, 24, 15, 19, 12, 8, 4, 0]
Test 7, RTL:
Move left by one word
"abc def hij AAA BBB WWW opr uvw xyz"[0, 7, 3, 12, 16, 20, 24, 31, 27]
Move right by one word
"abc def hij AAA BBB WWW opr uvw xyz"[35, 27, 31, 24, 20, 16, 12, 3, 7, 0]
Test 8, LTR:
Move right by one word
"abc def ZQB RIG uvw xyz"[0, 4, 8, 11, 16, 20]
Move left by one word
"abc def ZQB RIG uvw xyz"[23, 20, 16, 11, 8, 4, 0]
Test 9, RTL:
Move left by one word
"abc def ZQB RIG uvw xyz"[0, 3, 8, 12, 16, 19]
Move right by one word
"abc def ZQB RIG uvw xyz"[23, 19, 16, 12, 8, 3, 0]
Test 10, LTR:
Move right by one word
"aaa AAA bbb"[0, 4, 8]
Move left by one word
"aaa AAA bbb"[11, 8, 4, 0]
Test 11, RTL:
Move left by one word
"aaa AAA bbb"[0, 4, 8, 11]
Move right by one word
"aaa AAA bbb"[11, 8, 4, 0]
Test 12, LTR:
Move right by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[0, 7, 3, 12, 16, 20, 24, 31, 27]
Move left by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[35, 27, 31, 24, 20, 16, 12, 3, 7, 0]
Test 13, RTL:
Move left by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[0, 4, 8, 12, 19, 15, 24, 28, 32]
Move right by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[35, 32, 28, 24, 15, 19, 12, 8, 4, 0]
Test 14, LTR:
Move right by one word
"AAA BBB aaa bbb WWW DDD"[0, 3, 8, 12, 16, 19]
Move left by one word
"AAA BBB aaa bbb WWW DDD"[23, 19, 16, 12, 8, 3, 0]
Test 15, RTL:
Move left by one word
"AAA BBB aaa bbb WWW DDD"[0, 4, 8, 11, 16, 20]
Move right by one word
"AAA BBB aaa bbb WWW DDD"[23, 20, 16, 11, 8, 4, 0]
Test 16, LTR:
Move right by one word
"ZQB abc RIG"[0, 4, 8, 11]
Move left by one word
"ZQB abc RIG"[11, 8, 4, 0]
Test 17, RTL:
Move left by one word
"ZQB abc RIG"[0, 4, 8]
Move right by one word
"ZQB abc RIG"[11, 8, 4, 0]
Test 18, LTR:
Move right by one word
"abc def    hij opq"[0, 4, 8, 15]
Move left by one word
"abc def    hij opq"[18, 15, 8, 4, 0]
Test 19, LTR:
Move right by one word
"    abc    def    hij    opq    "[4, 8, 15, 22]
Move left by one word
"    abc    def    hij    opq    "[28, 22, 15, 8, 4]
Test 20, LTR:
Move right by one word
"    abc    ABW    def    "[4, 8, 15]
Move left by one word
"    abc    ABW    def    "[21, 15, 8, 4]
Test 21, LTR:
Move right by one word
"    abc    def    ABW    DDU    hij    opq    "[4, 8, 15, 21, 29, 36]
Move left by one word
"    abc    def    ABW    DDU    hij    opq    "[42, 36, 29, 21, 15, 8, 4]
Test 22, LTR:
Move right by one word
"    abc    def    hij    ABW    DSU    EJH    opq    rst    uvw    "[4, 8, 15, 22, 35, 28, 43, 50, 57]
Move left by one word
"    abc    def    hij    ABW    DSU    EJH    opq    rst    uvw    "[63, 57, 50, 43, 28, 35, 22, 15, 8, 4]
Test 23, LTR:
Move right by one word
"    ABW    DSU    HJH    FUX    "[4, 21, 14, 7]
Move left by one word
"    ABW    DSU    HJH    FUX    "[28, 7, 14, 21, 4]
Test 24, LTR:
Move right by one word
"    ABW    abc    DSU     "[4, 8, 15, 21]    FAIL expected: [4, 8, 15]
"    ABW    abc    DSU     "[15, 21]   FAIL expected to stay in the same position
"    ABW    abc    DSU     "[20, 21]   FAIL expected to stay in the same position
"    ABW    abc    DSU     "[19, 21]   FAIL expected to stay in the same position
Move left by one word
"    ABW    abc    DSU     "[21, 15, 8, 4]
Test 25, LTR:
Move right by one word
"    ABW    DSU    abc   def   HJH    FUX    "[4, 7, 15, 22, 28, 33]
Move left by one word
"    ABW    DSU    abc   def   HJH    FUX    "[40, 33, 28, 22, 15, 7, 4]
Test 26, LTR:
Move right by one word
"    ABW    DSU    HJH    abc   def   jih   FUX  FUX    YR[     "[4, 14, 7, 22, 29, 35, 41, 51, 46]
Move left by one word
"    ABW    DSU    HJH    abc   def   jih   FUX  FUX    YR[     "[58, 46, 51, 41, 35, 29, 22, 7, 14, 4]
Test 27, LTR:
Move right by one word
"ABW DSU    EJH FUX"[0, 14, 7, 3]
Move left by one word
"ABW DSU    EJH FUX"[18, 3, 7, 14, 0]
Test 28, LTR:
Move right by one word
"ABW DSU EJH    abc def hij"[0, 7, 3, 12, 19, 23]
Move left by one word
"ABW DSU EJH    abc def hij"[26, 23, 19, 12, 3, 7, 0]
Test 29, LTR:
Move right by one word
"abc def hij    ABW DSU EJH    opq rst uvw"[0, 4, 8, 12, 22, 18, 27, 34, 38]
Move left by one word
"abc def hij    ABW DSU EJH    opq rst uvw"[41, 38, 34, 27, 18, 22, 12, 8, 4, 0]
Test 30, RTL:
Move left by one word
"abc def    hij opq"[0, 14, 7, 3]
Move right by one word
"abc def    hij opq"[18, 3, 7, 14, 0]
Test 31, RTL:
Move left by one word
"    abc    def    hij    opq    "[4, 21, 14, 7]
Move right by one word
"    abc    def    hij    opq    "[28, 7, 14, 21, 4]
Test 32, RTL:
Move left by one word
"    abc    ABW    def    "[4, 8, 15, 21]    FAIL expected: [4, 8, 15]
"    abc    ABW    def    "[15, 21]   FAIL expected to stay in the same position
"    abc    ABW    def    "[20, 21]   FAIL expected to stay in the same position
"    abc    ABW    def    "[19, 21]   FAIL expected to stay in the same position
Move right by one word
"    abc    ABW    def    "[21, 15, 8, 4]
Test 33, RTL:
Move left by one word
"    abc    def    ABW    DDU    hij    opq    "[4, 7, 15, 22, 29, 35]
Move right by one word
"    abc    def    ABW    DDU    hij    opq    "[42, 35, 29, 22, 15, 7, 4]
Test 34, RTL:
Move left by one word
"    abc    def    hij    ABW    DSU    EJH    opq    rst    uvw    "[4, 14, 7, 22, 29, 36, 43, 56, 49]
Move right by one word
"    abc    def    hij    ABW    DSU    EJH    opq    rst    uvw    "[63, 49, 56, 43, 36, 29, 22, 7, 14, 4]
Test 35, RTL:
Move left by one word
"    ABW    DSU    HJH    FUX    "[4, 8, 15, 22]
Move right by one word
"    ABW    DSU    HJH    FUX    "[28, 22, 15, 8, 4]
Test 36, RTL:
Move left by one word
"    ABW    abc    DSU     "[4, 8, 15]
Move right by one word
"    ABW    abc    DSU     "[21, 15, 8, 4]
Test 37, RTL:
Move left by one word
"    ABW    DSU    abc   def   HJH    FUX    "[4, 8, 15, 21, 28, 34]
Move right by one word
"    ABW    DSU    abc   def   HJH    FUX    "[40, 34, 28, 21, 15, 8, 4]
Test 38, RTL:
Move left by one word
"    ABW    DSU    HJH    abc   def   jih   FUX  FUX    YR[     "[4, 8, 15, 22, 34, 28, 41, 47, 52]
Move right by one word
"    ABW    DSU    HJH    abc   def   jih   FUX  FUX    YR[     "[58, 52, 47, 41, 28, 34, 22, 15, 8, 4]
Test 39, LTR:
Move right by one word
"AAA "[0, 3]
Move left by one word
"BB"[2], "AAA "[3, 0]
Test 40, RTL:
Move left by one word
"AAA "[0, 4]
Move right by one word
"BB"[2], "AAA "[4, 0]
Test 41, LTR:
Move right by one word
"abc def "[0, 4, 8], "hij opq"[4], " rst uvw"[1, 5]
Move left by one word
" rst uvw"[8, 5, 1], "hij opq"[4], "abc def "[8, 4, 0]
Test 42, RTL:
Move left by one word
"abc def "[0], " rst uvw"[4], "hij opq"[3], "abc def "[7, 3]    FAIL expected: ["abc def "[ 0, ]" rst uvw"[ 4, ]"hij opq"[ 7,  3, ]"abc def "[ 7,  3]
" rst uvw"[4], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[3], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[2], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[1], "hij opq"[3]   FAIL expected "hij opq"[ 7]
Move right by one word
" rst uvw"[8], "abc def "[3, 7], "hij opq"[3], " rst uvw"[4], "abc def "[0]    FAIL expected: [" rst uvw"[ 8, ]"abc def "[ 3,  7, ]"hij opq"[ 3,  7, ]" rst uvw"[ 4, ]"abc def "[ 0]
"hij opq"[3], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[4], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[5], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[6], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
Test 43, RTL:
Move left by one word
"abc def "[0], " rst uvw"[4], "hij opq"[3], "abc def "[7, 3]    FAIL expected: ["abc def "[ 0, ]" rst uvw"[ 4, ]"hij opq"[ 7,  3, ]"abc def "[ 7,  3]
" rst uvw"[4], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[3], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[2], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[1], "hij opq"[3]   FAIL expected "hij opq"[ 7]
Move right by one word
" rst uvw"[8], "abc def "[3, 7], "hij opq"[3], " rst uvw"[4], "abc def "[0]    FAIL expected: [" rst uvw"[ 8, ]"abc def "[ 3,  7, ]"hij opq"[ 3,  7, ]" rst uvw"[ 4, ]"abc def "[ 0]
"hij opq"[3], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[4], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[5], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[6], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
Test 44, LTR:
Move right by one word
"abc def "[0, 4, 8], "hij opq"[4], " rst uvw"[1, 5]
Move left by one word
" rst uvw"[8, 5, 1], "hij opq"[4], "abc def "[8, 4, 0]
Test 45, RTL:
Move left by one word
"ABD DSU "[0, 4, 8], "EJH FUX"[4], "FFZ LIG"[4]
Move right by one word
"FFZ LIG"[7, 4], "EJH FUX"[4], "ABD DSU "[8, 4, 0]
Test 46, LTR:
Move right by one word
"ABD DSU "[0], "FFZ LIG"[3], "EJH FUX"[3], "ABD DSU "[7, 3]
Move left by one word
"FFZ LIG"[7], "ABD DSU "[3, 7], "EJH FUX"[3], "FFZ LIG"[3], "ABD DSU "[0]
Test 47, RTL:
Move left by one word
"ABD DSU "[0, 4, 8], "EJH FUX"[4], "FFZ LIG"[4]
Move right by one word
"FFZ LIG"[7, 4], "EJH FUX"[4], "ABD DSU "[8, 4, 0]
Test 48, LTR:
Move right by one word
"ABD DSU "[0], "FFZ LIG"[3], "EJH FUX"[3], "ABD DSU "[7, 3]
Move left by one word
"FFZ LIG"[7], "ABD DSU "[3, 7], "EJH FUX"[3], "FFZ LIG"[3], "ABD DSU "[0]
Test 49, RTL:
Move left by one word
"ABD DSU "[0, 4, 8], "abc def"[3], "FFZ LIG"[4]
Move right by one word
"FFZ LIG"[7, 4], "abc def"[3], "ABD DSU "[8, 4, 0]
Test 50, LTR:
Move right by one word
"ABD DSU "[0], "FFZ LIG"[3], "ABD DSU "[8], "abc def"[4], "ABD DSU "[7, 3]
Move left by one word
"FFZ LIG"[7], "ABD DSU "[3, 7], "abc def"[4], "ABD DSU "[8], "FFZ LIG"[3]    FAIL expected: ["FFZ LIG"[ 7, ]"ABD DSU "[ 3,  7, ]"abc def"[ 4, ]"ABD DSU "[ 8, ]"FFZ LIG"[ 3, ]"ABD DSU "[ 0]
"FFZ LIG"[3, 3]   FAIL expected "ABD DSU "[ 0]
"FFZ LIG"[4, 4]   FAIL expected "ABD DSU "[ 0]
"FFZ LIG"[5, 5]   FAIL expected "ABD DSU "[ 0]
"FFZ LIG"[6, 6]   FAIL expected "ABD DSU "[ 0]
Test 51, RTL:
Move left by one word
"ABD DSU "[0, 4, 8], "abc def"[3], "FFZ LIG"[4]
Move right by one word
"FFZ LIG"[7, 4], "abc def"[3], "ABD DSU "[8, 4, 0]
Test 52, LTR:
Move right by one word
"ABD DSU "[0, 3, 8], "abc def"[4], "FFZ LIG"[3]
Move left by one word
"FFZ LIG"[7, 3], "abc def"[4], "ABD DSU "[8, 3, 0]
Test 53, RTL:
Move left by one word
"ABD opq DSU "[0, 4, 8, 12], "abc AAA def"[8, 4, 3], "FFZ rst LIG"[4, 8]    FAIL expected: ["ABD opq DSU "[ 0,  4,  8,  12, ]"abc AAA def"[ 4,  3, ]"FFZ rst LIG"[ 4,  8]
"ABD opq DSU "[12], "abc AAA def"[8]   FAIL expected "abc AAA def"[ 4]
"abc AAA def"[10, 8]   FAIL expected "abc AAA def"[ 4]
"abc AAA def"[9, 8]   FAIL expected "abc AAA def"[ 4]
Move right by one word
"FFZ rst LIG"[11, 8, 4], "abc AAA def"[3, 4, 8], "ABD opq DSU "[12, 8, 4, 0]    FAIL expected: ["FFZ rst LIG"[ 11,  8,  4, ]"abc AAA def"[ 3,  4, ]"ABD opq DSU "[ 12,  8,  4,  0]
"abc AAA def"[4, 8]   FAIL expected "ABD opq DSU "[ 12]
Test 54, LTR:
Move right by one word
"ABD opq DSU "[0, 4], "abc AAA def"[8, 4], "ABD opq DSU "[12, 11], "FFZ rst LIG"[4, 8, 11]    FAIL expected: ["ABD opq DSU "[ 0,  4,  8, ]"abc AAA def"[ 8,  7, ]"ABD opq DSU "[ 12,  11, ]"FFZ rst LIG"[ 4,  8,  11]
"ABD opq DSU "[4], "abc AAA def"[8]   FAIL expected "ABD opq DSU "[ 8]
"ABD opq DSU "[5], "abc AAA def"[8]   FAIL expected "ABD opq DSU "[ 8]
"ABD opq DSU "[6], "abc AAA def"[8]   FAIL expected "ABD opq DSU "[ 8]
"ABD opq DSU "[7], "abc AAA def"[8]   FAIL expected "ABD opq DSU "[ 8]
"abc AAA def"[8, 4]   FAIL expected "abc AAA def"[ 7]
"abc AAA def"[9, 4]   FAIL expected "abc AAA def"[ 7]
"abc AAA def"[10, 4]   FAIL expected "abc AAA def"[ 7]
"abc AAA def"[11, 4]   FAIL expected "abc AAA def"[ 7]
"abc AAA def"[7, 4]   FAIL expected "ABD opq DSU "[ 12]
"abc AAA def"[6, 4]   FAIL expected "ABD opq DSU "[ 12]
"abc AAA def"[5, 4]   FAIL expected "ABD opq DSU "[ 12]
Move left by one word
"FFZ rst LIG"[11, 8, 4], "ABD opq DSU "[11, 12], "abc AAA def"[7, 8], "ABD opq DSU "[4, 0]    FAIL expected: ["FFZ rst LIG"[ 11,  8,  4, ]"ABD opq DSU "[ 11,  12, ]"abc AAA def"[ 7,  8, ]"ABD opq DSU "[ 8,  4,  0]
"abc AAA def"[8], "ABD opq DSU "[4]   FAIL expected "ABD opq DSU "[ 8]
"FFZ rst LIG"[1], "ABD opq DSU "[4]   FAIL expected "ABD opq DSU "[ 8]
"FFZ rst LIG"[2], "ABD opq DSU "[4]   FAIL expected "ABD opq DSU "[ 8]
Test 55, RTL:
Move left by one word
"ABD opq DSU "[0, 4, 8, 12], "abc AAA def"[4, 8], "FFZ rst LIG"[4, 8]
Move right by one word
"FFZ rst LIG"[11, 8, 4], "abc AAA def"[8, 4], "ABD opq DSU "[12, 8, 4, 0]
Test 56, LTR:
Move right by one word
"ABD opq DSU "[0, 4, 8, 12], "abc AAA def"[4, 8], "FFZ rst LIG"[4, 8, 11]
Move left by one word
"FFZ rst LIG"[11, 8, 4], "abc AAA def"[8, 4], "ABD opq DSU "[12, 8, 4, 0]
Test 57, LTR:
Move right by one word
"aaa "[0, 4], "bbb AAA "[4, 7]
Move left by one word
"FFZ"[3], "bbb AAA "[7, 4], "aaa "[4, 0]
Test 58, RTL:
Move left by one word
"ABD opq rst DSU "[0, 4, 7, 12, 16], "abc uvw AAA def lmn"[12, 8, 7, 3], "ABW hij xyz FXX"[4, 7, 12]    FAIL expected: ["ABD opq rst DSU "[ 0,  4,  7,  12,  16, ]"abc uvw AAA def lmn"[ 15,  8,  7,  3, ]"ABW hij xyz FXX"[ 4,  7,  12]
"ABD opq rst DSU "[16], "abc uvw AAA def lmn"[12]   FAIL expected "abc uvw AAA def lmn"[ 15]
"abc uvw AAA def lmn"[15, 12]   FAIL expected "abc uvw AAA def lmn"[ 8]
"abc uvw AAA def lmn"[14, 12]   FAIL expected "abc uvw AAA def lmn"[ 8]
"abc uvw AAA def lmn"[13, 12]   FAIL expected "abc uvw AAA def lmn"[ 8]
Move right by one word
"ABW hij xyz FXX"[15, 12, 7, 4], "abc uvw AAA def lmn"[3, 7, 8, 15], "ABD opq rst DSU "[16, 12, 7, 4, 0]
